+++ whoot/__init__.py
@@ -1,1 +1,4 @@
 __version__ = "0.0.2.dev0"
+
+
+from .extract_noise import clip_loud_segments
